Removable

Now it's mainly just IRegisterable and ISystem plus their implementations.

See merge request !58

+++ src/engine/model/weapons/DamageInfo.java
@@ -2,16 +2,16 @@ package engine.model.weapons;
 
 /**
  * A class to more easily pass around the information that results from a projectile exploding.
- * The class could be replaced by a struct with litle change in functionality.
+ * The class could be replaced by a struct with little change in functionality.
  * @author Weston
  *
  */
 public class DamageInfo {
-	private int myDamage;
+	private double myDamage;
 	private int myMoney;
 	private int myKills;
 	
-	public DamageInfo(int damageDealt, int moneyGained, int unitsKilled){
+	public DamageInfo(double damageDealt, int moneyGained, int unitsKilled){
 		myDamage = damageDealt;
 		myMoney = moneyGained;
 		myKills = unitsKilled;
@@ -22,7 +22,7 @@ public class DamageInfo {
 	}
 
 	public int getDamage(){
-		return myDamage;
+		return (int) myDamage;
 	}
 	public int getMoney(){
 		return myMoney;
